# openapi.yaml
openapi: 3.1.0
info:
  title: Redubber API
  description: '# Redubber API - AI-Powered Video Redubbing Platform


    The Redubber API provides comprehensive video redubbing capabilities powered by
    OpenAI''s

    Whisper, GPT-4, and TTS technologies. Transform videos into different languages
    while

    preserving the original speaker''s tone, emotion, and delivery style.


    ## Key Features


    - **Intelligent Voice Refinement**: AI-powered voice instruction generation for
    natural-sounding TTS

    - **Multi-Voice Testing**: Generate previews for all available voices simultaneously

    - **Smart Caching**: Avoid redundant TTS generations with intelligent caching

    - **Async Processing**: High-performance async TTS with 5x throughput improvement

    - **Project Management**: Organize videos, track progress, and manage redubbing
    workflows


    ## Workflow


    1. Create a project and upload videos

    2. Start redubbing task (transcribe, translate, generate audio)

    3. Use voice refinement to test different voices and instructions

    4. Select the best voice and apply to all segments

    5. Download the final redubbed video


    ## Rate Limits


    - OpenAI API limits apply to transcription, translation, and TTS operations

    - Recommended: 1-3 concurrent redubbing tasks per instance

    - TTS generation: Up to 100 concurrent requests (configurable)'
  contact:
    name: Redubber Development Team
...
    email: support@redubber.io
  license:
...
paths:
...
  /api/videos/projects/{project_id}/scan:
    post:
      tags:
      - videos
      summary: Trigger Scan
      description: "Trigger asynchronous file scan for a project.\n\nInitiates background\
        \ scanning of the project directory to detect\nand analyze video and subtitle\
        \ files. Returns immediately while\nscan runs in background.\n\nArgs:\n  \
        \  project_id: Project to scan.\n    background_tasks: FastAPI background\
        \ tasks manager.\n    db: DatabaseManager dependency.\n    scanner: FileScanner\
        \ dependency.\n\nReturns:\n    Scan trigger confirmation with status.\n\n\
        Raises:\n    HTTPException: 404 if project not found.\n    HTTPException:\
        \ 409 if scan is already running for this project."
      operationId: trigger_scan_api_videos_projects__project_id__scan_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTriggerResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/videos/projects/{project_id}/videos:
    get:
      tags:
      - videos
      summary: List Videos
      description: "List all analyzed videos for a project with pipeline status.\n\
        \nReturns comprehensive video metadata including audio streams,\nsubtitles,\
        \ and current pipeline processing status from the\nredubber_tmp directory.\n\
        \nArgs:\n    project_id: Project identifier.\n    db: DatabaseManager dependency.\n\
        \nReturns:\n    List of video analysis records with pipeline status.\n\nRaises:\n\
        \    HTTPException: 404 if project not found."
      operationId: list_videos_api_videos_projects__project_id__videos_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: integer
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VideoAnalysis'
                title: Response List Videos Api Videos Projects  Project Id  Videos
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AudioStream:
      properties:
        index:
          type: integer
          title: Index
          description: Stream index in video container
        language:
          type: string
          title: Language
          description: Language code (e.g., 'en', 'de')
        codec:
          type: string
          title: Codec
          description: Audio codec name (e.g., 'aac', 'mp3')
        channels:
          anyOf:
          - type: integer
          - type: string
          title: Channels
          description: Number of audio channels or descriptive string
        sample_rate:
          anyOf:
          - type: integer
          - type: string
          title: Sample Rate
          description: Audio sample rate in Hz or descriptive string
      type: object
      required:
      - index
      - language
      - codec
      - channels
      - sample_rate
      title: AudioStream
      description: Audio stream metadata from video file.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    PipelineStatusResponse:
      properties:
        progress:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Progress
          description: Progress percentage (0-100)
        current_stage:
          type: string
          title: Current Stage
          description: Current pipeline stage name
        is_complete:
          type: boolean
          title: Is Complete
          description: True if all pipeline stages complete
      type: object
      required:
      - progress
      - current_stage
      - is_complete
      title: PipelineStatusResponse
      description: Pipeline processing status for a video.
...
    ScanTriggerResponse:
      properties:
        project_id:
          type: integer
          title: Project Id
          description: Project ID being scanned
        status:
          type: string
          enum:
          - started
          - already_running
          title: Status
          description: Scan operation status
        message:
          type: string
          title: Message
          description: Human-readable status message
      type: object
      required:
      - project_id
      - status
      - message
      title: ScanTriggerResponse
      description: Response schema for triggering project scan.
...
    SubtitleInfo:
      properties:
        language:
          type: string
          title: Language
          description: Detected language code
        embedded:
          type: boolean
          title: Embedded
          description: True if subtitle is embedded in video container
        path:
          type: string
          title: Path
          description: File path for external subtitles
          default: ''
      type: object
      required:
      - language
      - embedded
      title: SubtitleInfo
      description: Subtitle file metadata.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    VideoAnalysis:
      properties:
        id:
          type: integer
          title: Id
          description: Unique video analysis record ID
        filename:
          type: string
          title: Filename
          description: Video filename without path
        path:
          type: string
          title: Path
          description: Absolute path to video file
        size_mb:
          type: number
          title: Size Mb
          description: File size in megabytes
        duration_seconds:
          type: number
          title: Duration Seconds
          description: Video duration in seconds
        audio_streams:
          items:
            $ref: '#/components/schemas/AudioStream'
          type: array
          title: Audio Streams
          description: List of detected audio streams
        subtitles:
          items:
            $ref: '#/components/schemas/SubtitleInfo'
          type: array
          title: Subtitles
          description: List of available subtitles
        pipeline_status:
          anyOf:
          - $ref: '#/components/schemas/PipelineStatusResponse'
          - type: 'null'
          description: Pipeline processing status (null if not started)
      type: object
      required:
      - id
      - filename
      - path
      - size_mb
      - duration_seconds
      - audio_streams
      - subtitles
      title: VideoAnalysis
      description: Comprehensive video file analysis with pipeline status.
...